+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [0ver](https://0ver.org).
 
 ## [Unreleased]
 
+## Added
+
+- By default, we will now trim input for `cipher` function from whitespaces and return carriages
+
 ## [0.1.4] - 2019-07-19
 
 ## Added